# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Car API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update Car Information",
                "operationId": "update-car",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car ID",
                        "name": "carId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Details for updating the car",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.CarApiRequestUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new car",
                "operationId": "add-car",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Owner ID",
                        "name": "ownerId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Car details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.CarApiRequestAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/car.CarApiResponseAdd"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/car.CarApiResponseAdd"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Car",
                "operationId": "delete-car",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car ID",
                        "name": "carId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    }
                }
            }
        },
        "/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get car information by registration number",
                "operationId": "getInfoByRegNum-car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Car registration number",
                        "name": "regNum",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/car.CarApiResponseGetInfo"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    }
                }
            }
        },
        "/info/filter": {
            "get": {
                "description": "Retrieve car information based on various filters and pagination settings.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieve filtered and paginated car information",
                "operationId": "getFilteredAndPaginatedInfo-car",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number for pagination (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "description": "Filter parameters for car information",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/car.CarApiRequestGetFilteredAndPaginatedInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/car.CarApiResponseGetFilteredAndPaginatedInfo"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/car.CarApiResponseGetFilteredAndPaginatedInfo"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "Add a new person to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new person",
                "operationId": "add-people",
                "parameters": [
                    {
                        "description": "Person data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/people.PeopleApiRequestAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/people.PeopleApiResponseAdd"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.defaultResponses"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.defaultResponses": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "car.CarApiRequestAdd": {
            "type": "object",
            "properties": {
                "mark": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "regNum": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "car.CarApiRequestGetFilteredAndPaginatedInfo": {
            "type": "object",
            "properties": {
                "mark": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "model": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "patronymic": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "regNum": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "surname": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "year": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "car.CarApiRequestUpdate": {
            "type": "object",
            "properties": {
                "ownerId": {
                    "type": "integer"
                },
                "regNum": {
                    "type": "string"
                }
            }
        },
        "car.CarApiResponseAdd": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "car.CarApiResponseGetFilteredAndPaginatedInfo": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/car.CarApiResponseGetInfo"
                    }
                },
                "perPage": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "car.CarApiResponseGetInfo": {
            "type": "object",
            "properties": {
                "mark": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/car.CarApiResponseGetInfoOwner"
                },
                "regNum": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "car.CarApiResponseGetInfoOwner": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "people.PeopleApiRequestAdd": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "people.PeopleApiResponseAdd": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}